Remove Internals folder
