+++ src/ReactGraph.Tests/GraphTests.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Linq;
-using ReactGraph.Internals.Graph;
 using Shouldly;
 using Xunit;
