+++ src/ReactGraph/Construction/DependencyDescriptor.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
-using ReactGraph.Internals.NodeInfo;
+using ReactGraph.NodeInfo;
 
-namespace ReactGraph.Internals.Construction
+namespace ReactGraph.Construction
 {
     abstract class DependencyDescriptor
     {
