+++ src/ReactGraph/NodeInfo/INodeInfo.cs
@@ -1,4 +1,4 @@
-namespace ReactGraph.Internals.NodeInfo
+namespace ReactGraph.NodeInfo
 {
     interface INodeInfo : IValueSource
     {
