+++ src/ReactGraph/NodeInfo/INodeInfo`1.cs
@@ -1,4 +1,4 @@
-namespace ReactGraph.Internals.NodeInfo
+namespace ReactGraph.NodeInfo
 {
     interface INodeInfo<T> : INodeInfo, IValueSource<T>
     {
