+++ src/ReactGraph/NodeInfo/IValueSource.cs
@@ -1,4 +1,4 @@
-namespace ReactGraph.Internals.NodeInfo
+namespace ReactGraph.NodeInfo
 {
     interface IValueSource
     {
