+++ src/ReactGraph/NodeInfo/IWritableNodeInfo.cs
@@ -1,4 +1,4 @@
-namespace ReactGraph.Internals.NodeInfo
+namespace ReactGraph.NodeInfo
 {
     interface IWritableNodeInfo<T> : INodeInfo<T>, IValueSink<T>
     {
