+++ src/ReactGraph/ReactGraph.csproj
@@ -37,39 +37,39 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DependencyEngine.cs" />
-    <Compile Include="Internals\Construction\DependencyDescriptor`1.cs" />
-    <Compile Include="Internals\Api\ExpressionDefinition.cs" />
+    <Compile Include="Construction\DependencyDescriptor`1.cs" />
+    <Compile Include="Api\ExpressionDefinition.cs" />
     <Compile Include="IExpressionDefinition.cs" />
-    <Compile Include="Internals\Construction\DependencyDescriptor.cs" />
-    <Compile Include="Internals\Construction\FormulaDescriptor.cs" />
-    <Compile Include="Internals\Construction\FormulaNullReferenceException.cs" />
-    <Compile Include="Internals\Construction\MemberDependencyDescriptor.cs" />
-    <Compile Include="Internals\Construction\NullCheckRewriter.cs" />
-    <Compile Include="Internals\NodeInfo\IMaybe.cs" />
-    <Compile Include="Internals\NodeInfo\INodeInfo`1.cs" />
-    <Compile Include="Internals\NodeInfo\IValueSource`1.cs" />
-    <Compile Include="Internals\NodeInfo\IWritableNodeInfo.cs" />
-    <Compile Include="Internals\NodeInfo\Maybe.cs" />
-    <Compile Include="Internals\Notification\INotificationStrategy.cs" />
-    <Compile Include="Internals\ExpressionStringBuilder.cs" />
-    <Compile Include="Internals\Graph\Edge.cs" />
-    <Compile Include="Internals\Graph\DirectedGraph.cs" />
-    <Compile Include="Internals\Construction\ExpressionParser.cs" />
-    <Compile Include="Internals\NodeInfo\IValueSink.cs" />
-    <Compile Include="Internals\NodeInfo\IValueSource.cs" />
-    <Compile Include="Internals\NodeInfo\WritableNodeInfo.cs" />
-    <Compile Include="Internals\NodeInfo\FormulaExpressionInfo.cs" />
-    <Compile Include="Internals\NodeInfo\INodeInfo.cs" />
+    <Compile Include="Construction\DependencyDescriptor.cs" />
+    <Compile Include="Construction\FormulaDescriptor.cs" />
+    <Compile Include="Construction\FormulaNullReferenceException.cs" />
+    <Compile Include="Construction\MemberDependencyDescriptor.cs" />
+    <Compile Include="Construction\NullCheckRewriter.cs" />
+    <Compile Include="NodeInfo\IMaybe.cs" />
+    <Compile Include="NodeInfo\INodeInfo`1.cs" />
+    <Compile Include="NodeInfo\IValueSource`1.cs" />
+    <Compile Include="NodeInfo\IWritableNodeInfo.cs" />
+    <Compile Include="NodeInfo\Maybe.cs" />
+    <Compile Include="Notification\INotificationStrategy.cs" />
+    <Compile Include="Construction\ExpressionStringBuilder.cs" />
+    <Compile Include="Graph\Edge.cs" />
+    <Compile Include="Graph\DirectedGraph.cs" />
+    <Compile Include="Construction\ExpressionParser.cs" />
+    <Compile Include="NodeInfo\IValueSink.cs" />
+    <Compile Include="NodeInfo\IValueSource.cs" />
+    <Compile Include="NodeInfo\WritableNodeInfo.cs" />
+    <Compile Include="NodeInfo\FormulaExpressionInfo.cs" />
+    <Compile Include="NodeInfo\INodeInfo.cs" />
     <Compile Include="IMemberDefinition.cs" />
-    <Compile Include="Internals\NodeInfo\NodeRepository.cs" />
-    <Compile Include="Internals\Notification\NotifyPropertyChangedStrategy.cs" />
+    <Compile Include="NodeInfo\NodeRepository.cs" />
+    <Compile Include="Notification\NotifyPropertyChangedStrategy.cs" />
     <Compile Include="Internals\Visualisation\DotVisualisation.cs" />
     <Compile Include="IVisualisation.cs" />
     <Compile Include="Internals\Visualisation\VertexVisualProperties.cs" />
     <Compile Include="Properties\Annotations.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Internals\Graph\Vertex.cs" />
-    <Compile Include="Internals\Api\MemberDefinition.cs" />
+    <Compile Include="Graph\Vertex.cs" />
+    <Compile Include="Api\MemberDefinition.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="TODO.txt" />
